Refactor 구조 수정 #94
