+++ src/main/java/com/cojac/storyteller/common/swagger/config/LoginSwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.cojac.storyteller.common.config.swagger;
+package com.cojac.storyteller.common.swagger.config;
 
 import io.swagger.v3.oas.models.PathItem;
 import io.swagger.v3.oas.models.media.Schema;
