+++ src/main/java/com/cojac/storyteller/common/swagger/config/SwaggerConfig.java
@@ -1,4 +1,4 @@
-package com.cojac.storyteller.common.config.swagger;
+package com.cojac.storyteller.common.swagger.config;
 
 import io.swagger.v3.oas.models.Components;
 import io.swagger.v3.oas.models.OpenAPI;
